+++ R/fairness_tensor.R
@@ -3,10 +3,9 @@
 #' @description
 #' A fairness tensor is a list of groupwise confusion matrices.
 #'
-#' @param object ([data.table()] | [PredictionClassif])\cr
-#'   A data.table with columns `truth` and `prediction` or a [PredictionClassif].
-#' @param task ([TaskClassif])\cr
-#'   A [TaskClassif]. Needs `col_role` `"pta"` to be set.
+#' @param object ([data.table()] | [PredictionClassif] | [ResampleResult])\cr
+#'   A data.table with columns `truth` and `prediction`,
+#'   a [PredictionClassif] or a [ResampleResult].
 #' @param ... `any`\cr
 #'   Currently not used.
 #' @return
@@ -16,10 +15,13 @@
 #' task = tsk("compas")
 #' prediction = lrn("classif.rpart")$train(task)$predict(task)
 #' fairness_tensor(prediction, task)
-fairness_tensor = function(object, task, ...) {
+fairness_tensor = function(object, ...) {
   UseMethod("fairness_tensor")
 }
 
+#' @rdname fairness_tensor
+#' @param task ([TaskClassif])\cr
+#'   A [TaskClassif]. Needs `col_role` `"pta"` to be set.
 #' @export
 fairness_tensor.data.table = function(object, task, ...) { # nolint
   assert_names(colnames(object), must.include = c("truth", "prediction"))
@@ -33,6 +35,7 @@ fairness_tensor.data.table = function(object, task, ...) { # nolint
   fairness_tensor(prd, task)
 }
 
+#' @rdname fairness_tensor
 #' @export
 fairness_tensor.PredictionClassif = function(object, task, ...) { # nolint
   assert_pta_task(task)
@@ -43,4 +46,11 @@ fairness_tensor.PredictionClassif = function(object, task, ...) { # nolint
   cols = c(task$backend$primary_key, task$col_roles$pta)
   data = task$data(cols = cols)
   map(split(data, by = cols[2L], keep.by = FALSE), function(x) get_confusion(x[[1L]]))
+}
+
+
+#' @rdname fairness_tensor
+#' @export
+fairness_tensor.ResampleResult = function(object, ...) { # nolint
+  fairness_tensor(object$prediction(), task = object$task, ...)
 }