+++ tests/testthat/test_write_files.R
@@ -1,6 +1,8 @@
 test_that("write_files", {
   skip_on_cran()
   tdir = tempdir()
+  tdir = paste0(tdir, "/report")
+  if (!dir.exists(tdir)) dir.create(tdir)
   unlink(list.files(tdir, full.names = TRUE), recursive = TRUE)
   object = list("one" = 1, "two" = 1)
   write_files(object, tdir)