+++ NEWS.md
@@ -1,10 +1,15 @@
+
 # mlr3fairness 0.2.0
 
 * Added 3 types of reports: `report_modelcard`, `report_datasheet`, `report_fairness`
 * Added several new fairness metrics
+<<<<<<< HEAD
 * Added 5 new learners (2 classification 3 regression) from package `fairml`.
   * classification: `classif.fairzlrm` | `classif.fairfgrrm`
   * regression:     `regr.fairnclm` | `regr.fairzlm` | `regr.fairfrrm`. 
+=======
+* Added `MeasureSubgroup` and `groupwise_metrics` that allow for inspecing dis-aggregated fairness metrics.
+>>>>>>> origin
 
 
 # mlr3fairness 0.1.0